+++ app/models/organization_summary.rb
@@ -4,6 +4,7 @@
 #
 #  id                :integer          primary key
 #  chapter_count     :integer
+#  country           :string
 #  deleted_at        :datetime
 #  group_count       :integer
 #  organization_mlid :string(3)
@@ -15,7 +16,7 @@
 class OrganizationSummary < ApplicationRecord
   include PgSearch::Model
 
-  pg_search_scope :search, against: [:organization_name], using: { tsearch: { prefix: true } }
+  pg_search_scope :search, against: [:organization_name, :country], using: { tsearch: { prefix: true } }
   self.primary_key = :id
 
   def readonly?