+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+- Filtered out deleted data when viewing group analytics
+- Upgraded Tailwind from V3 to V4
+
 ## 0.36.0 - Reworked Student MLIDs and Tags 
 - Added the ability to delete Tags
 - Added ability to search students by both organization and tag name